requirements.txt
annotated-types==0.6.0
anyio==3.7.1
autocommand==2.2.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
dnspython==2.4.2
docopt==0.6.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.104.0
gitdb==4.0.11
GitPython==1.0.2
h11==0.14.0
httptools==0.6.1
idna==3.4
inflect==7.0.0
install==1.3.5
jaraco.context==4.3.0
jaraco.env==1.0.0
jaraco.functools==3.9.0
jaraco.text==3.11.1
joe==0.0.7
jose==1.0.0
jwt==1.3.1
MarkupSafe==2.1.3
more-itertools==10.1.0
motor==3.3.1
oauthlib==3.2.2
package-name==0.1
packaging==23.2
passlib==1.7.4
path==16.7.1
Pillow==10.1.0
pip-run==12.4.0
platformdirs==3.11.0
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pymongo==4.5.0
python-dotenv==1.0.0
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starkbank-ecdsa==2.2.0
starlette==0.27.0
tweepy==4.14.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
watchfiles==0.21.0
websockets==12.0
